+++ sdk/src/androidTest/java/ly/count/android/sdk/ConnectionQueueTests.java
@@ -120,7 +120,8 @@ public class ConnectionQueueTests {
         freshConnQ.setExecutor(executor);
         assertSame(executor, freshConnQ.getExecutor());
     }
-
+/*
+//todo think about these tests
     @Test
     public void testCheckInternalState_nullAppKey() {
         connQ.checkInternalState(); // shouldn't throw
@@ -202,7 +203,7 @@ public class ConnectionQueueTests {
             // success!
         }
     }
-
+*/
     /*
     @Test
     public void testBeginSession() throws JSONException, UnsupportedEncodingException {
@@ -234,6 +235,7 @@ public class ConnectionQueueTests {
         }
     }*/
 
+    /*
     @Test
     public void testUpdateSession_checkInternalState() {
         try {
@@ -243,6 +245,7 @@ public class ConnectionQueueTests {
             // success!
         }
     }
+    */
 
     @Test
     public void testUpdateSession_zeroDuration() {
@@ -275,6 +278,7 @@ public class ConnectionQueueTests {
     //    assertEquals("60", queryParams.get("session_duration"));
     //}
 
+    /*
     @Test
     public void testEndSession_checkInternalState() {
         try {
@@ -284,6 +288,7 @@ public class ConnectionQueueTests {
             // success!
         }
     }
+    */
 
     /*
        @Test
@@ -352,6 +357,7 @@ public class ConnectionQueueTests {
            assertEquals("15", queryParams.get("session_duration"));
        }
        */
+    /*
     @Test
     public void testRecordEvents_checkInternalState() {
         try {
@@ -361,6 +367,7 @@ public class ConnectionQueueTests {
             // success!
         }
     }
+    */
 
     /*
     @Test
@@ -495,7 +502,7 @@ public class ConnectionQueueTests {
                         break;
                     case "sdk_version":
                         if (a == 0) {
-                            Assert.assertTrue(pair[1].equals("22.09.3"));
+                            Assert.assertTrue(pair[1].equals("22.09.4"));
                         } else if (a == 1) {
                             Assert.assertTrue(pair[1].equals("123sdf.v-213"));
                         }