+++ sdk/src/androidTest/java/ly/count/android/sdk/TestUtils.java
@@ -50,6 +50,21 @@ public class TestUtils {
     public static class Activity3 extends Activity {
     }
 
+    public static CountlyConfig createConfigurationConfig(boolean enableServerConfig, ImmediateRequestGenerator irGen) {
+        CountlyConfig cc = (new CountlyConfig((Application) ApplicationProvider.getApplicationContext(), commonAppKey, commonURL))
+            .setDeviceId(commonDeviceId)
+            .setLoggingEnabled(true)
+            .enableCrashReporting();
+
+        cc.immediateRequestGenerator = irGen;
+
+        if (enableServerConfig) {
+            cc.enableServerConfiguration();
+        }
+
+        return cc;
+    }
+
     public static CountlyConfig createConsentCountlyConfig(boolean requiresConsent, String[] givenConsent, ModuleBase testModuleListener, RequestQueueProvider rqp) {
         CountlyConfig cc = (new CountlyConfig((Application) ApplicationProvider.getApplicationContext(), commonAppKey, commonURL))
             .setDeviceId(commonDeviceId)