+++ sdk/src/main/java/ly/count/android/sdk/ConnectionQueue.java
@@ -24,6 +24,7 @@ package ly.count.android.sdk;
 import android.content.Context;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import java.time.OffsetDateTime;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -59,6 +60,7 @@ class ConnectionQueue implements RequestQueueProvider {
     protected ModuleRequestQueue moduleRequestQueue = null;//todo remove in the future
     protected DeviceInfo deviceInfo = null;//todo ?remove in the future?
     StorageProvider storageProvider;
+    ConfigurationProvider configProvider;
 
     void setBaseInfoProvider(BaseInfoProvider bip) {
         baseInfoProvider = bip;
@@ -122,22 +124,39 @@ class ConnectionQueue implements RequestQueueProvider {
      *
      * @throws IllegalStateException if context, app key, store, or server URL have not been set
      */
-    void checkInternalState() {
+    boolean checkInternalState() {
         if (context_ == null) {
-            throw new IllegalStateException("context has not been set");
+            if (L != null) {
+                L.e("[Connection Queue] context has not been set");
+            }
+            return false;
         }
         if (baseInfoProvider.getAppKey() == null || baseInfoProvider.getAppKey().length() == 0) {
-            throw new IllegalStateException("app key has not been set");
+            if (L != null) {
+                L.e("[Connection Queue] app key has not been set");
+            }
+            return false;
         }
         if (storageProvider == null) {
-            throw new IllegalStateException("countly storage provider has not been set");
+            if (L != null) {
+                L.e("[Connection Queue] countly storage provider has not been set");
+            }
+            return false;
         }
         if (baseInfoProvider.getServerURL() == null || !UtilsNetworking.isValidURL(baseInfoProvider.getServerURL())) {
-            throw new IllegalStateException("server URL is not valid");
+            if (L != null) {
+                L.e("[Connection Queue] server URL is not valid");
+            }
+            return false;
         }
         if (Countly.publicKeyPinCertificates != null && !baseInfoProvider.getServerURL().startsWith("https")) {
-            throw new IllegalStateException("server must start with https once you specified public keys");
+            if (L != null) {
+                L.e("[Connection Queue] server must start with https once you specified public keys");
+            }
+            return false;
         }
+
+        return true;
     }
 
     /**
@@ -146,7 +165,9 @@ class ConnectionQueue implements RequestQueueProvider {
      * @throws IllegalStateException if context, app key, store, or server URL have not been set
      */
     public void beginSession(boolean locationDisabled, @Nullable String locationCountryCode, @Nullable String locationCity, @Nullable String locationGpsCoordinates, @Nullable String locationIpAddress, @NonNull String preparedMetrics) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] beginSession");
 
         boolean dataAvailable = false;//will only send data if there is something valuable to send
@@ -181,7 +202,9 @@ class ConnectionQueue implements RequestQueueProvider {
      * @throws IllegalStateException if context, app key, store, or server URL have not been set
      */
     public void updateSession(final int duration) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] updateSession");
 
         if (duration > 0) {
@@ -201,14 +224,10 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void changeDeviceId(String deviceId, final int duration) {
-        checkInternalState();
-        L.d("[Connection Queue] changeDeviceId");
-
-        if (!consentProvider.anyConsentGiven()) {
-            L.d("[Connection Queue] request ignored, consent not given");
-            //no consent set, aborting
+        if (!checkInternalState()) {
             return;
         }
+        L.d("[Connection Queue] changeDeviceId");
 
         String data = prepareCommonRequestData();
 
@@ -224,11 +243,13 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void tokenSession(String token, Countly.CountlyMessagingMode mode, Countly.CountlyMessagingProvider provider) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] tokenSession");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.push)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'push' consent not given");
             return;
         }
 
@@ -265,7 +286,9 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void endSession(final int duration, String deviceIdOverride) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] endSession");
 
         boolean dataAvailable = false;//will only send data if there is something valuable to send
@@ -295,7 +318,9 @@ class ConnectionQueue implements RequestQueueProvider {
      * Send user location
      */
     public void sendLocation(boolean locationDisabled, String locationCountryCode, String locationCity, String locationGpsCoordinates, String locationIpAddress) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendLocation");
 
         String data = prepareCommonRequestData();
@@ -313,11 +338,13 @@ class ConnectionQueue implements RequestQueueProvider {
      * @throws java.lang.IllegalStateException if context, app key, store, or server URL have not been set
      */
     public void sendUserData(String userdata) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendUserData");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.users)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'user' consent not given");
             return;
         }
 
@@ -334,11 +361,13 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendIndirectAttribution(@NonNull String attributionObj) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendIndirectAttribution");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.attribution)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'attribution' consent not given");
             return;
         }
 
@@ -356,11 +385,13 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendDirectAttributionTest(@NonNull String attributionData) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendDirectAttributionTest");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.attribution)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'attribution' consent not given");
             return;
         }
 
@@ -378,11 +409,13 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendDirectAttributionLegacy(@NonNull String campaignID, @Nullable String userID) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendDirectAttributionLegacy");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.attribution)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'attribution' consent not given");
             return;
         }
 
@@ -412,11 +445,13 @@ class ConnectionQueue implements RequestQueueProvider {
      * @throws IllegalStateException if context, app key, store, or server URL have not been set
      */
     public void sendCrashReport(@NonNull final String crashData, final boolean nonFatalCrash) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendCrashReport");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.crashes)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'crashes' consent not given");
             return;
         }
 
@@ -437,7 +472,9 @@ class ConnectionQueue implements RequestQueueProvider {
      * @param requestData key value pair for direct request
      */
     public void sendDirectRequest(@NonNull final Map<String, String> requestData) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendDirectRequest");
 
         if (!consentProvider.anyConsentGiven()) {
@@ -467,7 +504,9 @@ class ConnectionQueue implements RequestQueueProvider {
      * @throws IllegalStateException if context, app key, store, or server URL have not been set
      */
     public void recordEvents(final String events) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendConsentChanges");
 
         ////////////////////////////////////////////////////
@@ -482,7 +521,9 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendConsentChanges(String formattedConsentChanges) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
         L.d("[Connection Queue] sendConsentChanges");
 
         final String data = prepareCommonRequestData()
@@ -494,12 +535,14 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendAPMCustomTrace(String key, Long durationMs, Long startMs, Long endMs, String customMetrics) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
 
         L.d("[Connection Queue] sendAPMCustomTrace");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.apm)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'apm' consent not given");
             return;
         }
 
@@ -519,12 +562,14 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendAPMNetworkTrace(String networkTraceKey, Long responseTimeMs, int responseCode, int requestPayloadSize, int responsePayloadSize, Long startMs, Long endMs) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
 
         L.d("[Connection Queue] sendAPMNetworkTrace");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.apm)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'apm' consent not given");
             return;
         }
 
@@ -545,7 +590,9 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendAPMAppStart(long durationMs, Long startMs, Long endMs) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
 
         L.d("[Connection Queue] sendAPMAppStart");
 
@@ -568,12 +615,14 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public void sendAPMScreenTime(boolean recordForegroundTime, long durationMs, Long startMs, Long endMs) {
-        checkInternalState();
+        if (!checkInternalState()) {
+            return;
+        }
 
         L.d("[Connection Queue] sendAPMScreenTime, recording foreground time: [" + recordForegroundTime + "]");
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.apm)) {
-            L.d("[Connection Queue] request ignored, consent not given");
+            L.d("[Connection Queue] request ignored, 'apm' consent not given");
             return;
         }
 
@@ -598,6 +647,15 @@ class ConnectionQueue implements RequestQueueProvider {
             + "&hour=" + instant.hour
             + "&dow=" + instant.dow
             + "&tz=" + deviceInfo.mp.getTimezoneOffset()
+            + "&sdk_version=" + Countly.sharedInstance().COUNTLY_SDK_VERSION_STRING
+            + "&sdk_name=" + Countly.sharedInstance().COUNTLY_SDK_NAME;
+    }
+
+    String prepareCommonRequestDataShort() {
+        UtilsTime.Instant instant = UtilsTime.getCurrentInstant();
+
+        return "app_key=" + UtilsNetworking.urlEncodeString(baseInfoProvider.getAppKey())
+            + "&timestamp=" + instant.timestampMs
             + "&sdk_version=" + Countly.sharedInstance().COUNTLY_SDK_VERSION_STRING
             + "&sdk_name=" + Countly.sharedInstance().COUNTLY_SDK_NAME;
     }
@@ -668,6 +726,14 @@ class ConnectionQueue implements RequestQueueProvider {
         return data;
     }
 
+    @Override
+    public String prepareServerConfigRequest() {
+        String data = prepareCommonRequestDataShort()
+            + "&method=sc"
+            + "&device_id=" + UtilsNetworking.urlEncodeString(deviceIdProvider_.getDeviceId());
+        return data;
+    }
+
     /**
      * Ensures that an executor has been created for ConnectionProcessor instances to be submitted to.
      */
@@ -699,7 +765,7 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public ConnectionProcessor createConnectionProcessor() {
-        return new ConnectionProcessor(baseInfoProvider.getServerURL(), storageProvider, deviceIdProvider_, sslContext_, requestHeaderCustomValues, L);
+        return new ConnectionProcessor(baseInfoProvider.getServerURL(), storageProvider, deviceIdProvider_, configProvider, sslContext_, requestHeaderCustomValues, L);
     }
 
     public boolean queueContainsTemporaryIdItems() {